+++ app/models/folio/session_attachment/base.rb
@@ -34,10 +34,7 @@ class Folio::SessionAttachment::Base < Folio::ApplicationRecord
                          optional: true,
                          touch: true
 
-  before_validation :set_file_mime_type
   after_save :pregenerate_thumbnails
-
-  alias_attribute :mime_type, :file_mime_type
 
   def to_h
     {
@@ -99,12 +96,6 @@ class Folio::SessionAttachment::Base < Folio::ApplicationRecord
   end
 
   private
-    def set_file_mime_type
-      return unless will_save_change_to_file_uid?
-      return unless file.present?
-      self.file_mime_type = file.mime_type
-    end
-
     def validate_type
       return errors.add(:type, :blank) if type.blank?
       return errors.add(:type, :invalid) if type.start_with?("Folio::")