+++ test/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_11_19_120347) do
+ActiveRecord::Schema.define(version: 2022_01_20_132205) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -215,7 +215,6 @@ ActiveRecord::Schema.define(version: 2021_11_19_120347) do
     t.integer "file_width"
     t.integer "file_height"
     t.bigint "file_size"
-    t.string "mime_type", limit: 255
     t.json "additional_data"
     t.json "file_metadata"
     t.string "hash_id"
@@ -224,6 +223,7 @@ ActiveRecord::Schema.define(version: 2021_11_19_120347) do
     t.integer "file_placements_size"
     t.string "file_name_for_search"
     t.boolean "sensitive_content", default: false
+    t.string "file_mime_type"
     t.index "to_tsvector('simple'::regconfig, folio_unaccent(COALESCE((author)::text, ''::text)))", name: "index_folio_files_on_by_author", using: :gin
     t.index "to_tsvector('simple'::regconfig, folio_unaccent(COALESCE((file_name_for_search)::text, ''::text)))", name: "index_folio_files_on_by_file_name_for_search", using: :gin
     t.index ["created_at"], name: "index_folio_files_on_created_at"